Changes to fix 64-bit Windows build errors and warnings.

+++ storage/spider/hs_client/config.cpp
@@ -263,7 +263,7 @@ parse_args(int argc, char **argv, config& conf)
     }
     if (!(param = new conf_param()))
       continue;
-    uint32 key_len = eq - arg;
+    uint32 key_len = (uint32)(eq - arg);
     uint32 val_len = strlen(eq + 1);
     if (
       param->key.reserve(key_len + 1) ||
